+++ odoo/local-src/procurement_sale/__openerp__.py
@@ -27,6 +27,7 @@
     'category': 'Stock Management',
     'depends': [
         'procurement',
+        'sale_stock',
         ],
     'data': [
     ],