+++ odoo/local-src/procurement_sale/models/stock.py
@@ -19,24 +19,28 @@
 #
 ##############################################################################
 
-{
-    'name': 'Stock Refill',
-    'version': '1.0',
-    'author': "BCIM",
-    'maintainer': 'Camptocamp',
-    'category': 'Stock Management',
-    'depends': [
-        'stock',
-        'stock_quant_bylocation',
-        ],
-    'data': [
-        'views/stock_location.xml',
-        'views/stock.xml',
-        'views/stock_reassort.xml',
-        'wizards/create_picking.xml',
-    ],
-    'installable': True,
-    'auto_install': False,
-    'license': 'AGPL-3',
-    'application': False,
-}
+from openerp import models, api
+
+
+class SaleOrderLine(models.Model):
+    _inherit = "sale.order.line"
+
+    @api.multi
+    def _action_procurement_create(self):
+        return super(SaleOrderLine, self.with_context(recount=True)). \
+            _action_procurement_create()
+
+
+class StockQuant(models.Model):
+    _inherit = 'stock.quant'
+
+    def quants_get_preferred_domain(self, cr, uid, qty, move, ops=False,
+                                    lot_id=False, domain=None,
+                                    preferred_domain_list=[], context=None):
+        if context and context.get('recount'):
+            remaining = (move.product_id.qty_available -
+                         move.product_id.outgoing_qty)
+            qty = min(qty, max(remaining, 0.0))
+        return super(StockQuant, self).quants_get_preferred_domain(
+            cr, uid, qty, move, ops=ops, lot_id=lot_id, domain=domain,
+            preferred_domain_list=[], context=context)